Add: ability to update todos. Improve framerate. Simplify diff animation. Split in multiple files.
