--- front/components/assistant/conversation/space/conversations/ProjectTodosPanel.tsx
@@ -1,1558 +1,0 @@
-import { AgentPicker } from "@app/components/assistant/AgentPicker";
-import { ConversationSidebarStatusDot } from "@app/components/assistant/conversation/ConversationSidebarStatusDot";
-import { ConfirmContext } from "@app/components/Confirm";
-import {
-  useSpaceConversations,
-  useSpaceConversationsSummary,
-} from "@app/hooks/conversations";
-import { useTodoDiffAnimations } from "@app/hooks/useTodoDiffAnimations";
-import { useAppRouter } from "@app/lib/platform";
-import {
-  useAgentConfigurations,
-  useUnifiedAgentConfigurations,
-} from "@app/lib/swr/assistants";
-import {
-  useBulkUpdateProjectTodoStatus,
-  useCleanDoneProjectTodos,
-  useCreateProjectTodo,
-  useDeleteProjectTodo,
-  useMarkProjectTodosRead,
-  useProjectTodos,
-  useStartProjectTodoConversation,
-  useUpdateProjectTodo,
-} from "@app/lib/swr/projects";
-import { useSpaceInfo } from "@app/lib/swr/spaces";
-import { timeAgoFrom } from "@app/lib/utils";
-import type { ConversationDotStatus } from "@app/lib/utils/conversation_dot_status";
-import { getConversationRoute } from "@app/lib/utils/router";
-import type { GetProjectTodosResponseBody } from "@app/pages/api/w/[wId]/spaces/[spaceId]/project_todos/index";
-import type { LightAgentConfigurationType } from "@app/types/assistant/agent";
-import {
-  compareAgentsForSort,
-  GLOBAL_AGENTS_SID,
-} from "@app/types/assistant/assistant";
-import type {
-  ProjectTodoActorType,
-  ProjectTodoAssigneeType,
-  ProjectTodoStatus,
-  ProjectTodoType,
-} from "@app/types/project_todo";
-import { assertNeverAndIgnore } from "@app/types/shared/utils/assert_never";
-import type { LightWorkspaceType } from "@app/types/user";
-import {
-  AnimatedText,
-  Avatar,
-  BookOpenIcon,
-  Button,
-  ChatBubbleLeftRightIcon,
-  Checkbox,
-  ChevronDownIcon,
-  ConfluenceLogo,
-  cn,
-  DriveLogo,
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuSearchbar,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-  GithubLogo,
-  Icon,
-  MicrosoftLogo,
-  NotionLogo,
-  PlayIcon,
-  PlusIcon,
-  RobotIcon,
-  SlackLogo,
-  Spinner,
-  TextArea,
-  Tooltip,
-  TrashIcon,
-  TypingAnimation,
-  UserGroupIcon,
-  UserIcon,
-  WindIcon,
-} from "@dust-tt/sparkle";
-import type React from "react";
-import {
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
-
-const SUMMARY_ITEM_TRANSITION_MS = 240;
-const DELETE_TODO_CONFIRM_PREVIEW_MAX_CHARS = 200;
-const NEW_MANUAL_TODO_MAX_CHARS = 256;
-
-// ── Metadata tooltip ──────────────────────────────────────────────────────────
-
-function formatActorLabel(
-  type: ProjectTodoActorType | null,
-  agentId: string | null,
-  agentNameById: Map<string, string>
-): string {
-  if (!type) {
-    return "someone";
-  }
-  switch (type) {
-    case "agent":
-      if (agentId === "butler") {
-        return "Dust";
-      }
-      const name = agentId ? agentNameById.get(agentId) : null;
-      return name ? `@${name}` : "an agent";
-    case "user":
-      return "a user";
-    default:
-      assertNeverAndIgnore(type);
-      return "someone";
-  }
-}
-
-function formatFriendlyDate(value: Date | string): string {
-  return `${timeAgoFrom(new Date(value).getTime(), { useLongFormat: true })} ago`;
-}
-
-interface TodoMetadataTooltipProps {
-  todo: ProjectTodoType;
-  agentNameById: Map<string, string>;
-  children: React.ReactElement;
-}
-
-function TodoMetadataTooltip({
-  todo,
-  agentNameById,
-  children,
-}: TodoMetadataTooltipProps) {
-  const creatorLabel = formatActorLabel(
-    todo.createdByType,
-    todo.createdByAgentConfigurationId,
-    agentNameById
-  );
-  const doneLabel = todo.markedAsDoneByType
-    ? formatActorLabel(
-        todo.markedAsDoneByType,
-        todo.markedAsDoneByAgentConfigurationId,
-        agentNameById
-      )
-    : null;
-
-  const isAssistantWorkInProgress =
-    !!todo.conversationId && todo.status === "in_progress";
-
-  const label = (
-    <div className="flex flex-col gap-1">
-      {isAssistantWorkInProgress && (
-        <div className="text-xs font-medium text-foreground dark:text-foreground-night">
-          An Agent is working on this todo.
-        </div>
-      )}
-      <div className="text-xs">
-        Created by {creatorLabel} · {formatFriendlyDate(todo.createdAt)}
-      </div>
-      {todo.doneAt && doneLabel && (
-        <div className="text-xs">
-          Done by {doneLabel} · {formatFriendlyDate(todo.doneAt)}
-        </div>
-      )}
-      {todo.actorRationale && (
-        <div className="max-w-xs text-xs italic opacity-80">
-          {todo.actorRationale}
-        </div>
-      )}
-    </div>
-  );
-
-  return <Tooltip label={label} tooltipTriggerAsChild trigger={children} />;
-}
-
-function useAgentNameById(
-  owner: LightWorkspaceType,
-  disabled?: boolean
-): Map<string, string> {
-  const { agentConfigurations } = useAgentConfigurations({
-    workspaceId: owner.sId,
-    agentsGetView: "list",
-    disabled,
-  });
-  return useMemo(() => {
-    const map = new Map<string, string>();
-    for (const a of agentConfigurations) {
-      map.set(a.sId, a.name);
-    }
-    return map;
-  }, [agentConfigurations]);
-}
-
-// ── Shared sub-components ─────────────────────────────────────────────────────
-
-function getSourceDisplay(source: ProjectTodoType["sources"][number]) {
-  const sourceIconByType: Record<
-    ProjectTodoType["sources"][number]["sourceType"],
-    React.ComponentType
-  > = {
-    project_conversation: ChatBubbleLeftRightIcon,
-    project_knowledge: BookOpenIcon,
-    slack: SlackLogo,
-    notion: NotionLogo,
-    gdrive: DriveLogo,
-    confluence: ConfluenceLogo,
-    github: GithubLogo,
-    microsoft: MicrosoftLogo,
-  };
-
-  const originalLabel = source.sourceTitle ?? source.sourceId;
-  const customLabel = source.sourceType === "slack" ? "Slack thread" : null;
-
-  return {
-    icon: sourceIconByType[source.sourceType],
-    label: customLabel ?? originalLabel,
-    originalLabel,
-    hasCustomLabel: customLabel !== null,
-  };
-}
-
-function TodoSources({
-  sources,
-  owner,
-  isDone,
-}: {
-  sources: ProjectTodoType["sources"];
-  owner: LightWorkspaceType;
-  isDone: boolean;
-}) {
-  const router = useAppRouter();
-
-  if (sources.length === 0) {
-    return null;
-  }
-
-  return (
-    <span
-      className={cn(
-        "text-xs",
-        isDone
-          ? "text-faint dark:text-faint-night line-through"
-          : "text-muted-foreground dark:text-muted-foreground-night"
-      )}
-    >
-      From{" "}
-      {sources.map((source, index) => (
-        <span key={`${source.sourceType}-${source.sourceId}`}>
-          {index > 0 && ", "}
-          <span
-            className={cn(
-              "relative inline-block",
-              isDone &&
-                "after:pointer-events-none after:absolute after:left-0 after:right-0 after:top-1/2 after:border-t after:border-current after:opacity-70"
-            )}
-          >
-            {(() => {
-              const { icon, label, originalLabel, hasCustomLabel } =
-                getSourceDisplay(source);
-
-              const trigger = (
-                <button
-                  type="button"
-                  className="underline hover:no-underline"
-                  onClick={(e) => {
-                    e.preventDefault();
-                    e.stopPropagation();
-
-                    if (!source.sourceUrl) {
-                      return;
-                    }
-
-                    try {
-                      const currentOrigin = window.location.origin;
-                      const targetUrl = new URL(
-                        source.sourceUrl,
-                        currentOrigin
-                      );
-
-                      if (targetUrl.origin === currentOrigin) {
-                        const internalPath = `${targetUrl.pathname}${targetUrl.search}${targetUrl.hash}`;
-                        void router.push(internalPath);
-                        return;
-                      }
-
-                      window.open(
-                        targetUrl.toString(),
-                        "_blank",
-                        "noopener,noreferrer"
-                      );
-                    } catch {
-                      void router.push(source.sourceUrl);
-                    }
-                  }}
-                >
-                  <Icon
-                    visual={icon}
-                    size="xs"
-                    className="mr-1 inline-block align-text-bottom opacity-70"
-                  />
-                  <span>{label}</span>
-                </button>
-              );
-
-              if (!hasCustomLabel) {
-                return trigger;
-              }
-
-              return <Tooltip label={originalLabel} trigger={trigger} />;
-            })()}
-          </span>
-        </span>
-      ))}
-    </span>
-  );
-}
-
-// ── Read-only panel (archived projects) ───────────────────────────────────────
-
-function ReadOnlyTodoItem({
-  todo,
-  owner,
-  agentNameById,
-}: {
-  todo: ProjectTodoType;
-  owner: LightWorkspaceType;
-  agentNameById: Map<string, string>;
-}) {
-  const isDone = todo.status === "done";
-
-  return (
-    <div className="flex items-start gap-3 overflow-hidden">
-      <div className="mt-0.5 shrink-0">
-        <Checkbox size="xs" checked={isDone} disabled />
-      </div>
-      <TodoMetadataTooltip todo={todo} agentNameById={agentNameById}>
-        <div className="flex flex-col gap-0.5">
-          <span
-            className={cn(
-              "select-text text-base min-h-6",
-              isDone
-                ? "text-faint dark:text-faint-night line-through"
-                : "text-foreground dark:text-foreground-night"
-            )}
-          >
-            {todo.text}
-          </span>
-          <div className="ml-1">
-            <TodoSources sources={todo.sources} owner={owner} isDone={isDone} />
-          </div>
-        </div>
-      </TodoMetadataTooltip>
-    </div>
-  );
-}
-
-function ReadOnlyProjectTodosPanel({
-  owner,
-  spaceId,
-}: {
-  owner: LightWorkspaceType;
-  spaceId: string;
-}) {
-  const { todos, isTodosLoading } = useProjectTodos({ owner, spaceId });
-  const agentNameById = useAgentNameById(owner);
-
-  return (
-    <div className="flex flex-col gap-3">
-      <h3 className="heading-2xl text-foreground dark:text-foreground-night">
-        To-dos
-      </h3>
-      {isTodosLoading ? (
-        <div className="flex justify-center py-4">
-          <Spinner size="sm" />
-        </div>
-      ) : (
-        <div className="flex flex-col gap-4">
-          <div className="flex flex-col">
-            {todos.map((todo) => (
-              <ReadOnlyTodoItem
-                key={todo.sId}
-                todo={todo}
-                owner={owner}
-                agentNameById={agentNameById}
-              />
-            ))}
-          </div>
-          {todos.length === 0 && (
-            <p className="text-base italic text-faint dark:text-faint-night">
-              You're all caught up!
-            </p>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ── Editable sub-components ───────────────────────────────────────────────────
-
-interface EditableTodoItemProps {
-  todo: ProjectTodoType;
-  viewerUserId: string | null;
-  onToggleDone: (todo: ProjectTodoType) => void;
-  onDelete: (todo: ProjectTodoType) => void | Promise<void>;
-  onStartWorking: (
-    todo: ProjectTodoType,
-    options?: { customMessage?: string; agentConfigurationId?: string }
-  ) => Promise<void>;
-  owner: LightWorkspaceType;
-  activeAgents: LightAgentConfigurationType[];
-  agentsLoading: boolean;
-  agentNameById: Map<string, string>;
-  isExiting: boolean;
-  isAdded: boolean;
-  isEntering: boolean;
-  isTyping: boolean;
-  isNewlyDone: boolean;
-  isStarting: boolean;
-}
-
-function EditableTodoItem({
-  todo,
-  viewerUserId,
-  onToggleDone,
-  onDelete,
-  onStartWorking,
-  owner,
-  activeAgents,
-  agentsLoading,
-  agentNameById,
-  isExiting,
-  isAdded,
-  isEntering,
-  isTyping,
-  isNewlyDone,
-  isStarting,
-}: EditableTodoItemProps) {
-  const router = useAppRouter();
-  const [startMenuOpen, setStartMenuOpen] = useState(false);
-  const [startCustomMessage, setStartCustomMessage] = useState("");
-  const [selectedStartAgent, setSelectedStartAgent] =
-    useState<LightAgentConfigurationType | null>(null);
-
-  const isDone = todo.status === "done";
-  const hasConversationLink =
-    (todo.status === "in_progress" || todo.status === "done") &&
-    !!todo.conversationId;
-  const conversationDotStatus: ConversationDotStatus =
-    todo.conversationSidebarStatus ?? "idle";
-  const isDoneWithoutConversation = isDone && !hasConversationLink;
-  const canEdit = viewerUserId !== null;
-  const showInProgressTextAnimation = todo.status === "in_progress";
-  const [isFlashing, setIsFlashing] = useState(isNewlyDone);
-
-  useEffect(() => {
-    if (!isNewlyDone) {
-      return;
-    }
-    const timeout = setTimeout(() => setIsFlashing(false), 1000);
-    return () => clearTimeout(timeout);
-  }, [isNewlyDone]);
-
-  const handleToggle = () => {
-    onToggleDone(todo);
-  };
-
-  const handleStartMenuOpenChange = (open: boolean) => {
-    setStartMenuOpen(open);
-    if (open) {
-      setStartCustomMessage("");
-      const defaultAgent =
-        activeAgents.find((a) => a.sId === GLOBAL_AGENTS_SID.DUST) ??
-        activeAgents[0] ??
-        null;
-      setSelectedStartAgent(defaultAgent);
-    }
-  };
-
-  const handleConfirmStart = async () => {
-    setStartMenuOpen(false);
-    const agentConfigurationId =
-      selectedStartAgent && selectedStartAgent.sId !== GLOBAL_AGENTS_SID.DUST
-        ? selectedStartAgent.sId
-        : undefined;
-    await onStartWorking(todo, {
-      customMessage: startCustomMessage.trim() || undefined,
-      agentConfigurationId,
-    });
-  };
-
-  return (
-    <div
-      className={cn(
-        "group/todo flex items-start gap-3 rounded-md px-1 py-1",
-        "transition-all duration-200 hover:bg-muted-background dark:hover:bg-muted-background-night",
-        isExiting
-          ? "max-h-0 overflow-hidden opacity-0"
-          : isAdded && !isEntering
-            ? "max-h-0 overflow-hidden opacity-0"
-            : "max-h-[1000px] opacity-100"
-      )}
-    >
-      <div className="mt-0.5 shrink-0">
-        <Checkbox
-          size="xs"
-          checked={isDone}
-          disabled={!canEdit}
-          isMutedAfterCheck
-          onCheckedChange={() => handleToggle()}
-        />
-      </div>
-      <TodoMetadataTooltip todo={todo} agentNameById={agentNameById}>
-        <div className="flex min-w-0 flex-1 flex-col gap-0.5">
-          <div className="cursor-default select-text text-left">
-            <span
-              className={cn(
-                "text-base min-h-6 transition-all duration-300",
-                isDone
-                  ? "text-faint dark:text-faint-night line-through"
-                  : "text-foreground dark:text-foreground-night",
-                isFlashing &&
-                  "rounded bg-warning-100/40 dark:bg-warning-100-night/30"
-              )}
-            >
-              {isTyping ? (
-                <TypingAnimation text={todo.text} duration={16} />
-              ) : showInProgressTextAnimation ? (
-                <AnimatedText variant="muted">{todo.text}</AnimatedText>
-              ) : (
-                todo.text
-              )}
-            </span>
-          </div>
-          <div className="ml-1">
-            <TodoSources sources={todo.sources} owner={owner} isDone={isDone} />
-          </div>
-        </div>
-      </TodoMetadataTooltip>
-      <div className="mt-0.5 flex shrink-0 items-center gap-1 opacity-100">
-        {hasConversationLink ? (
-          <Tooltip
-            label="Open to-do conversation"
-            trigger={
-              <span className="relative inline-flex shrink-0">
-                <Button
-                  icon={ChatBubbleLeftRightIcon}
-                  size="xs"
-                  variant="outline"
-                  onClick={() => {
-                    if (!todo.conversationId) {
-                      return;
-                    }
-                    void router.push(
-                      getConversationRoute(owner.sId, todo.conversationId),
-                      undefined,
-                      { shallow: true }
-                    );
-                  }}
-                />
-                <ConversationSidebarStatusDot
-                  status={conversationDotStatus}
-                  className="pointer-events-none absolute -right-0.5 -top-0.5 m-0 ring-2 ring-background dark:ring-background-night"
-                />
-              </span>
-            }
-          />
-        ) : (
-          canEdit &&
-          !hasConversationLink &&
-          (isDoneWithoutConversation ? (
-            <Tooltip
-              label="Reopen this to-do before starting work."
-              trigger={
-                <Button icon={PlayIcon} size="xs" variant="outline" disabled />
-              }
-            />
-          ) : (
-            <DropdownMenu
-              modal={false}
-              open={startMenuOpen}
-              onOpenChange={handleStartMenuOpenChange}
-            >
-              <DropdownMenuTrigger asChild>
-                <Button
-                  icon={PlayIcon}
-                  size="xs"
-                  variant="outline"
-                  isLoading={isStarting}
-                  disabled={isStarting}
-                  tooltip="Start working on to-do"
-                />
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="end" className="w-96 p-3">
-                <div className="flex flex-col gap-3">
-                  <TextArea
-                    id={`todo-start-msg-${todo.sId}`}
-                    aria-label="Additional instructions for the agent"
-                    placeholder="(optional) Add a custom message for the agent..."
-                    value={startCustomMessage}
-                    rows={4}
-                    onChange={(event: React.ChangeEvent<HTMLTextAreaElement>) =>
-                      setStartCustomMessage(event.target.value)
-                    }
-                  />
-                  <div className="flex items-center justify-between gap-2">
-                    <div className="w-[200px] min-w-0 shrink-0">
-                      <AgentPicker
-                        owner={owner}
-                        agents={activeAgents}
-                        disabled={agentsLoading}
-                        isLoading={agentsLoading}
-                        mountPortal
-                        showDropdownArrow
-                        showFooterButtons={false}
-                        side="bottom"
-                        size="xs"
-                        onItemClick={(agent) => setSelectedStartAgent(agent)}
-                        pickerButton={
-                          <Button
-                            variant="ghost-secondary"
-                            size="xs"
-                            isSelect
-                            icon={
-                              selectedStartAgent
-                                ? () => (
-                                    <Avatar
-                                      size="xxs"
-                                      visual={selectedStartAgent.pictureUrl}
-                                    />
-                                  )
-                                : RobotIcon
-                            }
-                            label={selectedStartAgent?.name ?? "Agent"}
-                            className="max-w-full min-w-0"
-                          />
-                        }
-                      />
-                    </div>
-                    <div className="w-[92px] shrink-0">
-                      <Button
-                        label="Go!"
-                        variant="highlight"
-                        size="sm"
-                        className="w-full"
-                        isLoading={isStarting}
-                        disabled={isStarting || !selectedStartAgent}
-                        onClick={() => void handleConfirmStart()}
-                      />
-                    </div>
-                  </div>
-                </div>
-              </DropdownMenuContent>
-            </DropdownMenu>
-          ))
-        )}
-        {canEdit && (
-          <div className="opacity-0 transition-opacity group-hover/todo:opacity-100">
-            <Tooltip
-              label="Delete to-do"
-              trigger={
-                <Button
-                  icon={TrashIcon}
-                  size="xs"
-                  variant="ghost"
-                  onClick={() => void onDelete(todo)}
-                />
-              }
-            />
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-export type TodoAssigneeScope = "mine" | "all" | "users";
-
-export interface TodoOwnerFilter {
-  assigneeScope: TodoAssigneeScope;
-  selectedUserSIds: string[];
-}
-
-function formatTodoScopeLabel({
-  scope,
-  selectedUserSIds,
-  usersBySId,
-  viewerUserId,
-}: {
-  scope: TodoAssigneeScope;
-  selectedUserSIds: Set<string>;
-  usersBySId: Map<string, ProjectTodoAssigneeType>;
-  viewerUserId: string | null;
-}) {
-  if (scope === "mine") {
-    return "Your to-dos";
-  }
-  if (scope === "all") {
-    return "Project's to-dos";
-  }
-
-  if (selectedUserSIds.size === 0) {
-    return "To-dos";
-  }
-
-  if (selectedUserSIds.size === 1) {
-    const [selectedUserSId] = selectedUserSIds;
-    const user = usersBySId.get(selectedUserSId);
-    if (!user) {
-      return "To-dos";
-    }
-
-    if (viewerUserId !== null && user.sId === viewerUserId) {
-      return "Your to-dos";
-    }
-
-    return `${user.fullName}'s to-dos`;
-  }
-
-  return `Selected to-dos (${selectedUserSIds.size})`;
-}
-
-function TodoAssigneeHeader({
-  user,
-  viewerUserId,
-}: {
-  user: ProjectTodoAssigneeType | null;
-  viewerUserId: string | null;
-}) {
-  const isYou = viewerUserId !== null && user?.sId === viewerUserId;
-
-  return (
-    <div className="mb-1 mt-2 flex items-center gap-2">
-      <Tooltip
-        label={user?.fullName ?? "Unknown user"}
-        trigger={
-          <Avatar
-            size="xxs"
-            visual={user?.image ?? "/static/humanavatar/anonymous.png"}
-          />
-        }
-      />
-      <span className="text-sm font-medium text-muted-foreground dark:text-muted-foreground-night">
-        {user?.fullName ?? "Unknown user"}
-        {isYou ? " (you)" : ""}
-      </span>
-    </div>
-  );
-}
-
-function EditableProjectTodosPanel({
-  owner,
-  spaceId,
-  todoOwnerFilter,
-  onTodoOwnerFilterChange,
-}: {
-  owner: LightWorkspaceType;
-  spaceId: string;
-  todoOwnerFilter: TodoOwnerFilter;
-  onTodoOwnerFilterChange: (value: TodoOwnerFilter) => void;
-}) {
-  const [assigneeSearch, setAssigneeSearch] = useState("");
-  const [isAssigneeMenuOpen, setIsAssigneeMenuOpen] = useState(false);
-  const {
-    todos,
-    users,
-    viewerUserId,
-    lastReadAt,
-    isTodosLoading,
-    mutateTodos,
-  } = useProjectTodos({
-    owner,
-    spaceId,
-  });
-  const agentNameById = useAgentNameById(owner);
-  const doUpdate = useUpdateProjectTodo({ owner, spaceId });
-  const doBulkUpdateStatus = useBulkUpdateProjectTodoStatus({ owner, spaceId });
-  const doDelete = useDeleteProjectTodo({ owner, spaceId });
-  const doStartConversation = useStartProjectTodoConversation({
-    owner,
-    spaceId,
-  });
-  const doCleanDone = useCleanDoneProjectTodos({ owner, spaceId });
-  const markRead = useMarkProjectTodosRead({ owner, spaceId });
-  const doCreateTodo = useCreateProjectTodo({ owner, spaceId });
-  const { spaceInfo, isSpaceInfoLoading } = useSpaceInfo({
-    workspaceId: owner.sId,
-    spaceId,
-  });
-  const confirm = useContext(ConfirmContext);
-
-  const { mutateConversations: mutateSpaceConversations } =
-    useSpaceConversations({
-      workspaceId: owner.sId,
-      spaceId,
-      options: { disabled: true },
-    });
-  const { mutate: mutateSpaceSummary } = useSpaceConversationsSummary({
-    workspaceId: owner.sId,
-    options: { disabled: true },
-  });
-
-  // Tracks todos being animated out during a clean operation.
-  const [pendingRemovalIds, setPendingRemovalIds] = useState<Set<string>>(
-    new Set()
-  );
-  const [isCleaning, setIsCleaning] = useState(false);
-  const [startingTodoIds, setStartingTodoIds] = useState<Set<string>>(
-    new Set()
-  );
-  const { agentConfigurations, isLoading: isAgentsLoading } =
-    useUnifiedAgentConfigurations({
-      workspaceId: owner.sId,
-      disabled: isTodosLoading,
-    });
-  const activeAgents = useMemo(() => {
-    const agents = agentConfigurations.filter((a) => a.status === "active");
-    agents.sort(compareAgentsForSort);
-    return agents;
-  }, [agentConfigurations]);
-
-  const projectMembers = useMemo(() => {
-    const members = spaceInfo?.members ?? [];
-    return [...members].sort((a, b) =>
-      a.fullName.localeCompare(b.fullName, undefined, { sensitivity: "base" })
-    );
-  }, [spaceInfo?.members]);
-
-  const defaultNewAssigneeSId = useMemo(() => {
-    if (projectMembers.length === 0) {
-      return null;
-    }
-    if (viewerUserId && projectMembers.some((m) => m.sId === viewerUserId)) {
-      return viewerUserId;
-    }
-    return projectMembers[0]!.sId;
-  }, [projectMembers, viewerUserId]);
-
-  const [newTodoText, setNewTodoText] = useState("");
-  const [newTodoAssigneeSId, setNewTodoAssigneeSId] = useState<string | null>(
-    null
-  );
-  const [addAssigneeSearch, setAddAssigneeSearch] = useState("");
-  const [isAddAssigneeMenuOpen, setIsAddAssigneeMenuOpen] = useState(false);
-  const [isAddingTodo, setIsAddingTodo] = useState(false);
-  const [isAddTodoComposerOpen, setIsAddTodoComposerOpen] = useState(false);
-  const addTodoInputRef = useRef<HTMLInputElement>(null);
-
-  const selectedNewAssigneeSId = newTodoAssigneeSId ?? defaultNewAssigneeSId;
-  const selectedNewAssigneeUser = useMemo(
-    () => projectMembers.find((m) => m.sId === selectedNewAssigneeSId) ?? null,
-    [projectMembers, selectedNewAssigneeSId]
-  );
-
-  const filteredAddAssigneeMembers = useMemo(() => {
-    const q = addAssigneeSearch.trim().toLowerCase();
-    if (!q) {
-      return projectMembers;
-    }
-    return projectMembers.filter((m) => m.fullName.toLowerCase().includes(q));
-  }, [addAssigneeSearch, projectMembers]);
-
-  // ── Diff animation state ────────────────────────────────────────────────────
-
-  // Frozen snapshot of lastReadAt taken on first successful load. undefined =
-  // not yet captured; null = first-ever visit; string = ISO timestamp.
-  // Initialized synchronously so new items start hidden from the very first
-  // render with data — prevents a layout shift on unchanged items caused by
-  // transitioning new items from visible→hidden after the first paint.
-  const [frozenLastReadAt, setFrozenLastReadAt] = useState<
-    string | null | undefined
-  >(() => (!isTodosLoading ? lastReadAt : undefined));
-
-  useEffect(() => {
-    if (!isTodosLoading && frozenLastReadAt === undefined) {
-      setFrozenLastReadAt(lastReadAt);
-    }
-  }, [isTodosLoading, frozenLastReadAt, lastReadAt]);
-
-  const { pendingAddedKeys, enteringKeys, typingKeys, doneFlashKeys } =
-    useTodoDiffAnimations({
-      ledgerScopeKey: `${owner.sId}:${spaceId}`,
-      todos,
-      frozenLastReadAt,
-      isTodosLoading,
-      markRead,
-    });
-
-  const usersBySId = useMemo(
-    () => new Map(users.map((user) => [user.sId, user])),
-    [users]
-  );
-  const selectedUserSIds = useMemo(
-    () => new Set(todoOwnerFilter.selectedUserSIds),
-    [todoOwnerFilter.selectedUserSIds]
-  );
-  const filteredUsers = useMemo(() => {
-    const normalizedSearch = assigneeSearch.trim().toLowerCase();
-    if (!normalizedSearch) {
-      return users;
-    }
-
-    return users.filter((user) =>
-      user.fullName.toLowerCase().includes(normalizedSearch)
-    );
-  }, [assigneeSearch, users]);
-  const todoScopeLabel = formatTodoScopeLabel({
-    scope: todoOwnerFilter.assigneeScope,
-    selectedUserSIds,
-    usersBySId,
-    viewerUserId,
-  });
-
-  const filteredTodos = useMemo(() => {
-    switch (todoOwnerFilter.assigneeScope) {
-      case "all":
-        return todos;
-      case "mine":
-        if (viewerUserId === null) {
-          return [];
-        }
-        return todos.filter((todo) => todo.user?.sId === viewerUserId);
-      case "users":
-        if (selectedUserSIds.size === 0) {
-          return todos;
-        }
-        return todos.filter(
-          (todo) => !!todo.user?.sId && selectedUserSIds.has(todo.user.sId)
-        );
-    }
-  }, [selectedUserSIds, todoOwnerFilter.assigneeScope, todos, viewerUserId]);
-
-  const hasDoneItems = filteredTodos.some((todo) => todo.status === "done");
-  const groupedTodosForAll = useMemo(() => {
-    const groups = new Map<
-      string,
-      { user: ProjectTodoAssigneeType | null; todos: ProjectTodoType[] }
-    >();
-
-    for (const todo of filteredTodos) {
-      const user = todo.user ?? null;
-      const key = user?.sId ?? `unknown-${todo.id}`;
-      const existing = groups.get(key);
-      if (existing) {
-        existing.todos.push(todo);
-      } else {
-        groups.set(key, { user, todos: [todo] });
-      }
-    }
-
-    return [...groups.values()].sort((a, b) => {
-      const aIsViewer = viewerUserId !== null && a.user?.sId === viewerUserId;
-      const bIsViewer = viewerUserId !== null && b.user?.sId === viewerUserId;
-      if (aIsViewer !== bIsViewer) {
-        return aIsViewer ? -1 : 1;
-      }
-      const aName = a.user?.fullName ?? "";
-      const bName = b.user?.fullName ?? "";
-      return aName.localeCompare(bName, undefined, { sensitivity: "base" });
-    });
-  }, [filteredTodos, viewerUserId]);
-
-  // Optimistically update a todo's status in the SWR cache and send the PATCH.
-  // On failure the cache is revalidated from the server.
-  const handleSetStatus = useCallback(
-    async (todo: ProjectTodoType, status: ProjectTodoStatus) => {
-      const optimistic = (prev: GetProjectTodosResponseBody | undefined) => ({
-        lastReadAt: prev?.lastReadAt ?? null,
-        viewerUserId: prev?.viewerUserId ?? viewerUserId,
-        todos: (prev?.todos ?? []).map((t) =>
-          t.sId === todo.sId ? { ...t, status } : t
-        ),
-      });
-
-      void mutateTodos(optimistic, { revalidate: false });
-
-      const result = await doUpdate(todo.sId, { status });
-      if (result.isErr()) {
-        // Revert: let SWR re-fetch the real server state.
-        void mutateTodos();
-      }
-    },
-    [doUpdate, mutateTodos, viewerUserId]
-  );
-
-  const handleToggleDone = useCallback(
-    (todo: ProjectTodoType) => {
-      if (todo.status === "done") {
-        void handleSetStatus(todo, "todo");
-      } else {
-        void handleSetStatus(todo, "done");
-      }
-    },
-    [handleSetStatus]
-  );
-
-  // Bulk set the status of every todo in a category in a single request.
-  // Optimistically updates the SWR cache, then revalidates on failure.
-  const _handleSetStatusForSection = useCallback(
-    async (status: ProjectTodoStatus) => {
-      const targetIds = todos
-        .filter((t) => t.status !== status)
-        .map((t) => t.sId);
-
-      if (targetIds.length === 0) {
-        return;
-      }
-
-      const targetIdSet = new Set(targetIds);
-      const optimistic = (prev: GetProjectTodosResponseBody | undefined) => ({
-        lastReadAt: prev?.lastReadAt ?? null,
-        viewerUserId: prev?.viewerUserId ?? viewerUserId,
-        todos: (prev?.todos ?? []).map((t) =>
-          targetIdSet.has(t.sId) ? { ...t, status } : t
-        ),
-      });
-
-      void mutateTodos(optimistic, { revalidate: false });
-
-      const result = await doBulkUpdateStatus(targetIds, status);
-      if (result.isErr()) {
-        void mutateTodos();
-      }
-    },
-    [doBulkUpdateStatus, mutateTodos, todos, viewerUserId]
-  );
-
-  const handleClean = useCallback(async () => {
-    setIsCleaning(true);
-
-    // Optimistically hide done items to trigger exit animations.
-    const doneSIds = new Set(
-      todos.filter((t) => t.status === "done").map((t) => t.sId)
-    );
-    setPendingRemovalIds(doneSIds);
-
-    const result = await doCleanDone();
-
-    if (result.isOk()) {
-      // Wait for exit animations to finish, then refresh the server data.
-      // pendingRemovalIds is cleared only after mutateTodos resolves so items
-      // don't briefly reappear from the stale SWR cache and re-trigger the
-      // exit animation.
-      setTimeout(async () => {
-        await mutateTodos();
-        setPendingRemovalIds(new Set());
-        setIsCleaning(false);
-      }, SUMMARY_ITEM_TRANSITION_MS);
-    } else {
-      // Revert on failure.
-      setPendingRemovalIds(new Set());
-      setIsCleaning(false);
-    }
-  }, [doCleanDone, todos, mutateTodos]);
-
-  const handleDelete = useCallback(
-    async (todo: ProjectTodoType) => {
-      const result = await doDelete(todo.sId);
-      if (result.isOk()) {
-        void mutateTodos();
-      }
-    },
-    [doDelete, mutateTodos]
-  );
-
-  const requestDelete = useCallback(
-    async (todo: ProjectTodoType) => {
-      const preview =
-        todo.text.length > DELETE_TODO_CONFIRM_PREVIEW_MAX_CHARS
-          ? `${todo.text.slice(0, DELETE_TODO_CONFIRM_PREVIEW_MAX_CHARS)}…`
-          : todo.text;
-      const confirmed = await confirm({
-        title: "Delete to-do?",
-        message: `“${preview}”`,
-        validateLabel: "Delete",
-        validateVariant: "warning",
-      });
-      if (!confirmed) {
-        return;
-      }
-      void handleDelete(todo);
-    },
-    [confirm, handleDelete]
-  );
-
-  const handleAddManualTodo = useCallback(async () => {
-    const text = newTodoText.trim();
-    if (!text || !selectedNewAssigneeSId || isAddingTodo) {
-      return;
-    }
-    setIsAddingTodo(true);
-    const result = await doCreateTodo({
-      text,
-      assigneeUserId: selectedNewAssigneeSId,
-    });
-    setIsAddingTodo(false);
-    if (result.isOk()) {
-      setNewTodoText("");
-      const created: ProjectTodoType = {
-        ...result.value,
-        conversationId: null,
-        conversationSidebarStatus: null,
-      };
-      await mutateTodos(
-        (prev: GetProjectTodosResponseBody | undefined) => ({
-          lastReadAt: prev?.lastReadAt ?? null,
-          viewerUserId: prev?.viewerUserId ?? null,
-          todos: [created, ...(prev?.todos ?? [])],
-        }),
-        { revalidate: true }
-      );
-      // After SWR updates, restore focus — rAF waits until layout/paint so the input isn’t skipped.
-      requestAnimationFrame(() => {
-        requestAnimationFrame(() => {
-          addTodoInputRef.current?.focus();
-        });
-      });
-    }
-  }, [
-    doCreateTodo,
-    isAddingTodo,
-    mutateTodos,
-    newTodoText,
-    selectedNewAssigneeSId,
-  ]);
-
-  const closeAddTodoComposer = useCallback(() => {
-    setIsAddAssigneeMenuOpen(false);
-    setNewTodoText("");
-    setNewTodoAssigneeSId(null);
-    setIsAddTodoComposerOpen(false);
-  }, []);
-
-  useEffect(() => {
-    if (isAddTodoComposerOpen) {
-      queueMicrotask(() => addTodoInputRef.current?.focus());
-    }
-  }, [isAddTodoComposerOpen]);
-
-  const handleStartWorking = useCallback(
-    async (
-      todo: ProjectTodoType,
-      options?: { customMessage?: string; agentConfigurationId?: string }
-    ) => {
-      setStartingTodoIds((prev) => new Set([...prev, todo.sId]));
-      const result = await doStartConversation(todo.sId, {
-        customMessage: options?.customMessage,
-        agentConfigurationId: options?.agentConfigurationId,
-      });
-      if (result.isOk()) {
-        const { conversationId } = result.value;
-        // Reflect the new todo state (conversationId set) immediately.
-        // Only patch conversationId — the server-side toJSON doesn't rehydrate
-        // sources, so replacing the whole todo would transiently drop them.
-        void mutateTodos(
-          (prev: GetProjectTodosResponseBody | undefined) => ({
-            lastReadAt: prev?.lastReadAt ?? null,
-            viewerUserId: prev?.viewerUserId ?? viewerUserId,
-            todos: (prev?.todos ?? []).map((t) =>
-              t.sId === todo.sId
-                ? {
-                    ...t,
-                    status: "in_progress",
-                    doneAt: null,
-                    markedAsDoneByType: null,
-                    markedAsDoneByAgentConfigurationId: null,
-                    conversationId,
-                    conversationSidebarStatus: "idle",
-                  }
-                : t
-            ),
-          }),
-          { revalidate: false }
-        );
-        // Revalidate the conversations list below so the new conversation
-        // appears right away.
-        void mutateSpaceConversations();
-        void mutateSpaceSummary();
-        void mutateTodos();
-      }
-      setStartingTodoIds((prev) => {
-        const next = new Set(prev);
-        next.delete(todo.sId);
-        return next;
-      });
-    },
-    [
-      doStartConversation,
-      mutateTodos,
-      mutateSpaceConversations,
-      mutateSpaceSummary,
-      viewerUserId,
-    ]
-  );
-
-  return (
-    <div className="flex flex-col gap-3">
-      {/* Header */}
-      <div className="inline-flex items-center gap-2">
-        <DropdownMenu
-          modal={false}
-          open={isAssigneeMenuOpen}
-          onOpenChange={(open) => {
-            setIsAssigneeMenuOpen(open);
-            if (open) {
-              setAssigneeSearch("");
-            }
-          }}
-        >
-          <DropdownMenuTrigger asChild>
-            <button
-              type="button"
-              className="inline-flex items-center gap-1 rounded-md px-1 py-0.5 hover:bg-muted/40 dark:hover:bg-muted-night/40"
-            >
-              <h3 className="heading-2xl text-foreground dark:text-foreground-night">
-                {todoScopeLabel}
-              </h3>
-              <Icon
-                visual={ChevronDownIcon}
-                size="sm"
-                className="text-muted-foreground dark:text-muted-foreground-night"
-              />
-            </button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent
-            className="z-[1000] w-80 shadow-2xl ring-1 ring-border/60"
-            align="start"
-          >
-            <DropdownMenuSearchbar
-              autoFocus
-              name="assignee-filter"
-              placeholder="Search members"
-              value={assigneeSearch}
-              onChange={setAssigneeSearch}
-            />
-            <DropdownMenuSeparator />
-            <DropdownMenuCheckboxItem
-              icon={UserIcon}
-              label="Your to-dos"
-              checked={todoOwnerFilter.assigneeScope === "mine"}
-              onClick={() => {
-                onTodoOwnerFilterChange({
-                  assigneeScope: "mine",
-                  selectedUserSIds: [],
-                });
-                setIsAssigneeMenuOpen(false);
-              }}
-              onSelect={(event) => {
-                event.preventDefault();
-              }}
-            />
-            <DropdownMenuCheckboxItem
-              icon={UserGroupIcon}
-              label="Project's to-dos"
-              checked={todoOwnerFilter.assigneeScope === "all"}
-              onClick={() => {
-                onTodoOwnerFilterChange({
-                  assigneeScope: "all",
-                  selectedUserSIds: [],
-                });
-                setIsAssigneeMenuOpen(false);
-              }}
-              onSelect={(event) => {
-                event.preventDefault();
-              }}
-            />
-            <DropdownMenuSeparator />
-            <div className="max-h-64 overflow-auto">
-              {filteredUsers.length > 0 ? (
-                filteredUsers.map((user) => (
-                  <DropdownMenuCheckboxItem
-                    key={`todo-assignee-filter-${user.sId}`}
-                    icon={() => (
-                      <Avatar
-                        size="xxs"
-                        visual={
-                          user.image ?? "/static/humanavatar/anonymous.png"
-                        }
-                      />
-                    )}
-                    label={`${user.fullName}${viewerUserId === user.sId ? " (you)" : ""}`}
-                    checked={
-                      todoOwnerFilter.assigneeScope === "users" &&
-                      selectedUserSIds.has(user.sId)
-                    }
-                    onClick={() => {
-                      const next = new Set(selectedUserSIds);
-                      if (next.has(user.sId)) {
-                        next.delete(user.sId);
-                      } else {
-                        next.add(user.sId);
-                      }
-                      onTodoOwnerFilterChange({
-                        assigneeScope: next.size === 0 ? "all" : "users",
-                        selectedUserSIds: [...next],
-                      });
-                    }}
-                    onSelect={(event) => {
-                      event.preventDefault();
-                    }}
-                  />
-                ))
-              ) : (
-                <div className="px-3 py-2 text-sm text-muted-foreground">
-                  No members found
-                </div>
-              )}
-            </div>
-          </DropdownMenuContent>
-        </DropdownMenu>
-        <div className="flex-1" />
-        {hasDoneItems && (
-          <Button
-            size="xs"
-            variant="outline"
-            icon={WindIcon}
-            label="Clean"
-            tooltip="Hide all done to-dos"
-            onClick={handleClean}
-            disabled={isCleaning}
-          />
-        )}
-      </div>
-
-      {/* Manual add: discreet until opened; one row when expanded */}
-      {isSpaceInfoLoading ? (
-        <div className="flex h-7 items-center">
-          <Spinner size="sm" />
-        </div>
-      ) : projectMembers.length === 0 ? (
-        <p className="text-sm text-muted-foreground dark:text-muted-foreground-night">
-          No project members available to assign.
-        </p>
-      ) : isAddTodoComposerOpen ? (
-        <div className="flex h-9 w-full min-w-0 items-center gap-1.5 rounded-lg border border-border/60 bg-background/80 px-1.5 py-0 shadow-sm dark:border-border-night/60 dark:bg-background-night/50">
-          <DropdownMenu
-            modal={false}
-            open={isAddAssigneeMenuOpen}
-            onOpenChange={(open) => {
-              setIsAddAssigneeMenuOpen(open);
-              if (open) {
-                setAddAssigneeSearch("");
-              }
-            }}
-          >
-            <DropdownMenuTrigger asChild>
-              <button
-                type="button"
-                className="inline-flex h-7 w-7 shrink-0 items-center justify-center rounded-md border border-border/80 bg-muted-background/50 text-foreground hover:bg-muted-background focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-highlight/30 dark:border-border-night/80 dark:bg-muted-background-night/50 dark:text-foreground-night dark:hover:bg-muted-background-night/80 dark:focus-visible:ring-highlight-night/40"
-                disabled={isAddingTodo}
-                title={
-                  selectedNewAssigneeUser
-                    ? `Assign to ${selectedNewAssigneeUser.fullName}${viewerUserId === selectedNewAssigneeUser.sId ? " (you)" : ""} — click to change`
-                    : "Choose assignee"
-                }
-                aria-label={
-                  selectedNewAssigneeUser
-                    ? `Assign to ${selectedNewAssigneeUser.fullName}${viewerUserId === selectedNewAssigneeUser.sId ? " (you)" : ""}, open menu to change`
-                    : "Choose assignee"
-                }
-              >
-                <Avatar
-                  size="xxs"
-                  visual={
-                    selectedNewAssigneeUser?.image ??
-                    "/static/humanavatar/anonymous.png"
-                  }
-                />
-              </button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent
-              className="z-[1000] w-80 shadow-2xl ring-1 ring-border/60"
-              align="start"
-            >
-              <DropdownMenuSearchbar
-                autoFocus
-                name="add-todo-assignee"
-                placeholder="Search members"
-                value={addAssigneeSearch}
-                onChange={setAddAssigneeSearch}
-              />
-              <DropdownMenuSeparator />
-              <div className="max-h-64 overflow-auto">
-                {filteredAddAssigneeMembers.length > 0 ? (
-                  filteredAddAssigneeMembers.map((member) => (
-                    <DropdownMenuCheckboxItem
-                      key={`add-todo-assignee-${member.sId}`}
-                      icon={() => (
-                        <Avatar
-                          size="xxs"
-                          visual={
-                            member.image ?? "/static/humanavatar/anonymous.png"
-                          }
-                        />
-                      )}
-                      label={`${member.fullName}${viewerUserId === member.sId ? " (you)" : ""}`}
-                      checked={selectedNewAssigneeSId === member.sId}
-                      onClick={() => {
-                        setNewTodoAssigneeSId(member.sId);
-                        setIsAddAssigneeMenuOpen(false);
-                      }}
-                      onSelect={(event) => {
-                        event.preventDefault();
-                      }}
-                    />
-                  ))
-                ) : (
-                  <div className="px-3 py-2 text-sm text-muted-foreground">
-                    No members found
-                  </div>
-                )}
-              </div>
-            </DropdownMenuContent>
-          </DropdownMenu>
-          <input
-            ref={addTodoInputRef}
-            type="text"
-            name="new-manual-project-todo"
-            aria-label="New to-do"
-            autoComplete="off"
-            maxLength={NEW_MANUAL_TODO_MAX_CHARS}
-            placeholder="Add a to-do — Enter to add · Esc to cancel"
-            value={newTodoText}
-            disabled={isAddingTodo}
-            className={cn(
-              "min-h-0 min-w-0 flex-1 border-0 bg-transparent py-0 text-base leading-6 text-foreground",
-              "shadow-none [box-shadow:none]",
-              "outline-none ring-0 ring-offset-0",
-              "focus:shadow-none focus:[box-shadow:none] focus:outline-none focus:ring-0 focus:ring-offset-0",
-              "focus:!ring-0 focus:!ring-offset-0",
-              "focus-visible:shadow-none focus-visible:outline-none focus-visible:ring-0 focus-visible:ring-offset-0",
-              "focus-visible:!ring-0",
-              "placeholder:text-muted-foreground",
-              "dark:text-foreground-night dark:placeholder:text-muted-foreground-night",
-              "disabled:opacity-60"
-            )}
-            onChange={(e) => setNewTodoText(e.target.value)}
-            onKeyDown={(e) => {
-              if (e.key === "Escape") {
-                e.preventDefault();
-                closeAddTodoComposer();
-                return;
-              }
-              if (e.key === "Enter") {
-                e.preventDefault();
-                void handleAddManualTodo();
-              }
-            }}
-          />
-          <Tooltip
-            label="Add to-do"
-            trigger={
-              <Button
-                size="xs"
-                variant="highlight"
-                icon={PlusIcon}
-                isLoading={isAddingTodo}
-                disabled={
-                  isAddingTodo || !newTodoText.trim() || !selectedNewAssigneeSId
-                }
-                onClick={() => void handleAddManualTodo()}
-              />
-            }
-          />
-        </div>
-      ) : (
-        <button
-          type="button"
-          onClick={() => setIsAddTodoComposerOpen(true)}
-          className="flex w-full min-h-9 items-center gap-2 rounded-md px-1.5 py-1.5 text-left text-sm text-muted-foreground transition-colors hover:bg-muted-background/70 dark:text-muted-foreground-night dark:hover:bg-muted-background-night/70"
-        >
-          <Icon visual={PlusIcon} size="xs" className="shrink-0 opacity-80" />
-          <span>Add a to-do</span>
-        </button>
-      )}
-
-      {/* Body */}
-      {isTodosLoading || frozenLastReadAt === undefined ? (
-        <div className="flex justify-center py-4">
-          <Spinner size="sm" />
-        </div>
-      ) : (
-        <div className="flex flex-col gap-4">
-          {/* Todo items */}
-          <div className="flex flex-col">
-            {groupedTodosForAll.map((group) => (
-              <div
-                key={group.user?.sId ?? `unknown-${group.todos[0]?.id}`}
-                className="mb-4 last:mb-0"
-              >
-                <TodoAssigneeHeader
-                  user={group.user}
-                  viewerUserId={viewerUserId}
-                />
-                <div className="ml-4 flex flex-col">
-                  {group.todos.map((todo) => (
-                    <EditableTodoItem
-                      key={todo.sId}
-                      todo={todo}
-                      viewerUserId={viewerUserId}
-                      onToggleDone={handleToggleDone}
-                      onDelete={requestDelete}
-                      onStartWorking={handleStartWorking}
-                      owner={owner}
-                      activeAgents={activeAgents}
-                      agentsLoading={isAgentsLoading}
-                      agentNameById={agentNameById}
-                      isExiting={pendingRemovalIds.has(todo.sId)}
-                      isAdded={pendingAddedKeys.has(todo.sId)}
-                      isEntering={enteringKeys.has(todo.sId)}
-                      isTyping={typingKeys.has(todo.sId)}
-                      isNewlyDone={doneFlashKeys.has(todo.sId)}
-                      isStarting={startingTodoIds.has(todo.sId)}
-                    />
-                  ))}
-                </div>
-              </div>
-            ))}
-          </div>
-
-          {/* Empty state */}
-          {filteredTodos.length === 0 && (
-            <p className="text-base italic text-faint dark:text-faint-night">
-              You're all caught up!
-            </p>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ── Main component ─────────────────────────────────────────────────────────────
-
-interface ProjectTodosPanelProps {
-  owner: LightWorkspaceType;
-  spaceId: string;
-  isReadOnly: boolean;
-  todoOwnerFilter: TodoOwnerFilter;
-  onTodoOwnerFilterChange: (value: TodoOwnerFilter) => void;
-}
-
-export function ProjectTodosPanel({
-  owner,
-  spaceId,
-  isReadOnly,
-  todoOwnerFilter,
-  onTodoOwnerFilterChange,
-}: ProjectTodosPanelProps) {
-  if (isReadOnly) {
-    return <ReadOnlyProjectTodosPanel owner={owner} spaceId={spaceId} />;
-  }
-
-  return (
-    <EditableProjectTodosPanel
-      owner={owner}
-      spaceId={spaceId}
-      todoOwnerFilter={todoOwnerFilter}
-      onTodoOwnerFilterChange={onTodoOwnerFilterChange}
-    />
-  );
-}
