+++ front/hooks/useTodoDiffAnimations.ts
@@ -1,7 +1,7 @@
 import type { ProjectTodoType } from "@app/types/project_todo";
 import { useEffect, useRef, useState } from "react";
 
-const SUMMARY_ITEM_TRANSITION_MS = 240;
+const DONE_FLASH_DURATION_MS = 1000;
 const TODO_ANIMATION_LEDGER_STORAGE_KEY = "project_todos_animation_ledger_v1";
 const TODO_ANIMATION_LEDGER_MAX_KEYS = 1000;
 
@@ -81,9 +81,7 @@ function addAnimatedTodoEventKeys(keys: Iterable<string>) {
 }
 
 interface TodoDiffAnimationState {
-  pendingAddedKeys: Set<string>;
-  enteringKeys: Set<string>;
-  typingKeys: Set<string>;
+  newItemKeys: Set<string>;
   doneFlashKeys: Set<string>;
 }
 
@@ -102,17 +100,13 @@ export function useTodoDiffAnimations({
   isTodosLoading,
   markRead,
 }: UseTodoDiffAnimationsArgs): TodoDiffAnimationState {
-  const [pendingAddedKeys, setPendingAddedKeys] = useState<Set<string>>(
-    new Set()
-  );
-  const [enteringKeys, setEnteringKeys] = useState<Set<string>>(new Set());
-  const [enteredKeys, setEnteredKeys] = useState<Set<string>>(new Set());
-  const [typingKeys, setTypingKeys] = useState<Set<string>>(new Set());
+  const [newItemKeys, setNewItemKeys] = useState<Set<string>>(new Set());
+  const [seenNewKeys, setSeenNewKeys] = useState<Set<string>>(new Set());
   const [doneFlashKeys, setDoneFlashKeys] = useState<Set<string>>(new Set());
-  const startRaf1Ref = useRef<number | null>(null);
-  const startRaf2Ref = useRef<number | null>(null);
-  const cleanupTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const inFlightAddedKeysRef = useRef<Set<string>>(new Set());
+  const inFlightNewKeysRef = useRef<Set<string>>(new Set());
+  const doneFlashTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(
+    null
+  );
 
   useEffect(() => {
     hydrateAnimatedTodoEventKeysFromSessionStorage();
@@ -138,8 +132,8 @@ export function useTodoDiffAnimations({
       if (createdAtMs > cutoffMs) {
         const eventKey = `${ledgerScopeKey}:added:${todo.sId}:${todo.createdAt}`;
         if (
-          !enteredKeys.has(todo.sId) &&
-          !inFlightAddedKeysRef.current.has(todo.sId) &&
+          !seenNewKeys.has(todo.sId) &&
+          !inFlightNewKeysRef.current.has(todo.sId) &&
           !animatedTodoEventKeys.has(eventKey)
         ) {
           addedTodoIds.add(todo.sId);
@@ -164,89 +158,50 @@ export function useTodoDiffAnimations({
       return;
     }
 
-    // Persist immediately so refocus/remount doesn't replay the same events.
     addAnimatedTodoEventKeys([...addedEventKeys, ...newlyDoneEventKeys]);
 
-    // Mark read immediately so navigating away/back during animation
-    // doesn't cause the same items to re-animate on next mount.
     void markRead();
 
     if (addedTodoIds.size > 0) {
-      // Keep these keys marked in-flight until cleanup completes; clearing them
-      // early lets focus/revalidation passes replay "added" enter animations.
-      inFlightAddedKeysRef.current = new Set(addedTodoIds);
-      setPendingAddedKeys(new Set(addedTodoIds));
-      setTypingKeys(new Set(addedTodoIds));
-    }
-    setDoneFlashKeys((prev) => new Set([...prev, ...newlyDoneTodoIds]));
-
-    // If a revalidation triggers another pass while an animation is in-flight,
-    // reschedule from the latest keys instead of letting React's effect cleanup
-    // cancel the previous run and leave items collapsed.
-    if (startRaf1Ref.current !== null) {
-      cancelAnimationFrame(startRaf1Ref.current);
-      startRaf1Ref.current = null;
-    }
-    if (startRaf2Ref.current !== null) {
-      cancelAnimationFrame(startRaf2Ref.current);
-      startRaf2Ref.current = null;
-    }
-    if (cleanupTimeoutRef.current !== null) {
-      clearTimeout(cleanupTimeoutRef.current);
-      cleanupTimeoutRef.current = null;
+      inFlightNewKeysRef.current = new Set([
+        ...inFlightNewKeysRef.current,
+        ...addedTodoIds,
+      ]);
+      setNewItemKeys((prev) => new Set([...prev, ...addedTodoIds]));
+      setSeenNewKeys((prev) => new Set([...prev, ...addedTodoIds]));
     }
 
-    // Double-RAF: wait for the browser to paint the initial hidden state of
-    // new items (isAdded && !isEntering -> max-h-0 opacity-0) before triggering
-    // the entering animation. setTimeout(0) could fire before the first paint,
-    // causing items to flash visible before animating in.
-    startRaf1Ref.current = requestAnimationFrame(() => {
-      startRaf2Ref.current = requestAnimationFrame(() => {
-        setEnteringKeys(new Set(addedTodoIds));
-        startRaf1Ref.current = null;
-        startRaf2Ref.current = null;
-      });
-    });
+    if (newlyDoneTodoIds.size > 0) {
+      setDoneFlashKeys((prev) => new Set([...prev, ...newlyDoneTodoIds]));
 
-    cleanupTimeoutRef.current = setTimeout(() => {
-      setEnteringKeys(new Set());
-      setPendingAddedKeys(new Set());
-      setEnteredKeys((prev) => new Set([...prev, ...addedTodoIds]));
-      inFlightAddedKeysRef.current = new Set();
-      cleanupTimeoutRef.current = null;
-    }, SUMMARY_ITEM_TRANSITION_MS);
+      if (doneFlashTimeoutRef.current !== null) {
+        clearTimeout(doneFlashTimeoutRef.current);
+      }
+      doneFlashTimeoutRef.current = setTimeout(() => {
+        setDoneFlashKeys(new Set());
+        doneFlashTimeoutRef.current = null;
+      }, DONE_FLASH_DURATION_MS);
+    }
   }, [
     frozenLastReadAt,
     isTodosLoading,
     ledgerScopeKey,
     markRead,
     todos,
-    enteredKeys,
+    seenNewKeys,
   ]);
 
-  // Cleanup only on unmount.
   useEffect(() => {
     return () => {
-      const hadInFlightAnimation = inFlightAddedKeysRef.current.size > 0;
+      if (doneFlashTimeoutRef.current !== null) {
+        clearTimeout(doneFlashTimeoutRef.current);
+      }
 
-      if (startRaf1Ref.current !== null) {
-        cancelAnimationFrame(startRaf1Ref.current);
-      }
-      if (startRaf2Ref.current !== null) {
-        cancelAnimationFrame(startRaf2Ref.current);
-      }
-      if (cleanupTimeoutRef.current !== null) {
-        clearTimeout(cleanupTimeoutRef.current);
-      }
-      inFlightAddedKeysRef.current = new Set();
-
-      // If the user navigates away before the animation cleanup timeout runs,
-      // persist the read marker so the same items don't re-animate on remount.
-      if (hadInFlightAnimation) {
+      if (inFlightNewKeysRef.current.size > 0) {
         void markRead();
       }
     };
   }, [markRead]);
 
-  return { pendingAddedKeys, enteringKeys, typingKeys, doneFlashKeys };
+  return { newItemKeys, doneFlashKeys };
 }
